FEAT: wordpress start.sh

+++ srcs/requirements/wordpress/tools/start.sh
@@ -5,19 +5,24 @@ mkdir -p /var/www/html
 
 cd /var/www/html
 
-rm -rf *
+if [ "$(pwd)" = "/var/www/html" ]; then
+    rm -rf *
+fi
 
-curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar 
+curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar || { echo "Failed to download wp-cli"; exit 1; }
 
 chmod +x wp-cli.phar 
 
 mv wp-cli.phar /usr/local/bin/wp
+
 
 wp core download --allow-root
 
 echo "moving file"
 
 mv /var/www/html/wp-config-sample.php /var/www/html/wp-config.php
+
+ls /var/www/html
 
 mv /wp-config.php /var/www/html/wp-config.php
 
@@ -31,6 +36,6 @@ wp core install --url=$DOMAIN_NAME --title=$WP_TITLE --admin_user=$WP_ADMIN_USER
 
 wp user create $WP_USER $WP_USER_EMAIL --role=author --user_pass=$WP_USER_PASSWORD --allow-root
 
-sed -i 's/listen = /run/php/php7.4-fpm.sock/listen = 9000/g' /etc/php/7.4/fpm/pool.d/www.conf
+sed -i 's/listen = \/run\/php\/php7.4-fpm.sock/listen = 9000/g' /etc/php/7.4/fpm/pool.d/www.conf
 
 /usr/sbin/php-fpm7.4 -F
